+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 
 ## [1.0.2] - 2020-06-25
 ### Changed
+- configuration.py.TEMPLATE:  Changed format of configuration file for Mongo connection.
 - Documentation updates.